+++ mobile/android/geckoview/src/main/java/org/mozilla/geckoview/WebRequestError.java
@@ -11,9 +11,14 @@ import org.mozilla.gecko.annotation.WrapForJNI;
 import android.annotation.SuppressLint;
 import android.support.annotation.AnyThread;
 import android.support.annotation.IntDef;
+import android.support.annotation.Nullable;
 
+import java.io.ByteArrayInputStream;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
+import java.security.cert.CertificateException;
+import java.security.cert.CertificateFactory;
+import java.security.cert.X509Certificate;
 
 /**
  * WebRequestError is simply a container for error codes and categories used by
@@ -218,14 +223,31 @@ public class WebRequestError extends Exception {
     public final int category;
 
     /**
+     * The server certificate used. This can be useful if the error code is
+     * is e.g. {@link #ERROR_SECURITY_BAD_CERT}.
+     */
+    public final @Nullable X509Certificate certificate;
+
+    /**
      * Construct a new WebRequestError with the specified code and category.
      * @param code An error code, e.g. {@link #ERROR_MALFORMED_URI}
      * @param category An error category, e.g. {@link #ERROR_CATEGORY_URI}
      */
     public WebRequestError(final @Error int code, final @ErrorCategory int category) {
+        this(code, category, null);
+    }
+
+    /**
+     * Construct a new WebRequestError with the specified code and category.
+     * @param code An error code, e.g. {@link #ERROR_MALFORMED_URI}
+     * @param category An error category, e.g. {@link #ERROR_CATEGORY_URI}
+     * @param certificate The X509Certificate server certificate used, if applicable.
+     */
+    public WebRequestError(final @Error int code, final @ErrorCategory int category, final X509Certificate certificate) {
         super(String.format("Request failed, error=0x%x, category=0x%x", code, category));
         this.code = code;
         this.category = category;
+        this.certificate = certificate;
     }
 
     @Override
@@ -236,6 +258,7 @@ public class WebRequestError extends Exception {
 
         final WebRequestError otherError = (WebRequestError)other;
 
+        // We don't compare the certificate here because it's almost never what you want.
         return otherError.code == this.code &&
                 otherError.category == this.category;
     }
@@ -248,10 +271,21 @@ public class WebRequestError extends Exception {
     @WrapForJNI
     /* package */ static WebRequestError fromGeckoError(final long geckoError,
                                                         final int geckoErrorModule,
-                                                        final int geckoErrorClass) {
+                                                        final int geckoErrorClass,
+                                                        final byte[] certificateBytes) {
         int code = convertGeckoError(geckoError, geckoErrorModule, geckoErrorClass);
         int category = getErrorCategory(geckoErrorModule, code);
-        return new WebRequestError(code, category);
+        X509Certificate certificate = null;
+        if (certificateBytes != null) {
+            try {
+                final CertificateFactory factory = CertificateFactory.getInstance("X.509");
+                certificate = (X509Certificate) factory.generateCertificate(new ByteArrayInputStream(certificateBytes));
+            } catch (CertificateException e) {
+                throw new IllegalArgumentException("Unable to parse DER certificate");
+            }
+        }
+
+        return new WebRequestError(code, category, certificate);
     }
 
     @SuppressLint("WrongConstant")